clean unnecessary code
