+++ app/src/androidTest/java/com/github/factotum_sdp/factotum/LoginMainActivityTest.kt
@@ -1,0 +1,31 @@
+package com.github.factotum_sdp.factotum
+
+import androidx.test.espresso.Espresso.onView
+import androidx.test.espresso.action.ViewActions.*
+import androidx.test.espresso.assertion.ViewAssertions.matches
+import androidx.test.espresso.intent.Intents
+import androidx.test.espresso.intent.Intents.intended
+import androidx.test.espresso.matcher.ViewMatchers.withId
+import androidx.test.espresso.intent.matcher.IntentMatchers.hasComponent
+import androidx.test.espresso.intent.matcher.IntentMatchers.hasExtra
+import androidx.test.espresso.matcher.ViewMatchers.withText
+import androidx.test.ext.junit.rules.ActivityScenarioRule
+import androidx.test.ext.junit.runners.AndroidJUnit4
+
+import org.junit.Rule
+import org.junit.Test
+import org.junit.runner.RunWith
+
+import org.hamcrest.core.AllOf.allOf
+
+
+@RunWith(AndroidJUnit4::class)
+class LoginMainActivityTest {
+
+    private val userName = "Carl"
+
+    @get:Rule
+    var testRule = ActivityScenarioRule(
+        LoginMainActivity::class.java
+    )
+}
