--- app/src/main/java/com/github/factotum_sdp/factotum/GreetingActivity.kt
@@ -1,19 +1,0 @@
-package com.github.factotum_sdp.factotum
-
-import android.os.Bundle
-import android.widget.TextView
-import androidx.appcompat.app.AppCompatActivity
-
-class GreetingActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        setContentView(R.layout.activity_greeting)
-        val plainText: TextView = findViewById(R.id.greetingMessage)
-        plainText.text =
-        buildString {
-            append("Hello ")
-            append(intent.getStringExtra("userName"))
-        }
-    }
-}
