+++ app/src/main/java/com/github/factotum_sdp/factotum/LoginMainActivity.kt
@@ -1,0 +1,15 @@
+package com.github.factotum_sdp.factotum
+
+import android.content.Intent
+import android.os.Bundle
+import android.widget.Button
+import android.widget.EditText
+import androidx.appcompat.app.AppCompatActivity
+
+class LoginMainActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+    }
+}
